pbs_backend/requirements.txt
Django==3.1.4
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
